Call `mergeConfigs` on `createOfflineMiddleware` and `enhanceReducer`
