+++ src/index.js
@@ -1,31 +1,13 @@
 // @flow
 /* global $Shape */
 import { applyMiddleware, compose } from 'redux';
-import type { Config } from './types';
-import { createOfflineMiddleware } from './middleware';
-import { enhanceReducer } from './updater';
-import { applyDefaults } from './config';
+
 import { networkStatusChanged } from './actions';
-import offlineActionTracker from './offlineActionTracker';
+import mergeConfigs from './mergeConfigs';
+import { createOfflineMiddleware } from './middleware';
+import type { Config } from './types';
+import { enhanceReducer } from './updater';
 
-// @TODO: Take createStore as config?
-const warnIfNotReduxAction = (config: $Shape<Config>, key: string) => {
-  const maybeAction = config[key];
-
-  const isNotReduxAction =
-    maybeAction === null ||
-    typeof maybeAction !== 'object' ||
-    typeof maybeAction.type !== 'string' ||
-    maybeAction.type === '';
-
-  if (isNotReduxAction && console.warn) {
-    const msg =
-      `${key} must be a proper redux action, ` +
-      `i.e. it must be an object and have a non-empty string type. ` +
-      `Instead you provided: ${JSON.stringify(maybeAction, null, 2)}`;
-    console.warn(msg);
-  }
-};
 
 export const offline = (userConfig: $Shape<Config> = {}) => (
   createStore: any
@@ -86,24 +68,6 @@ export const createEnhanceStore = (userConfig: $Shape<Config> = {}) => {
     return configStore(createStore(reducer, preloadedState, enhancer), config);
   };
 };
-
-function mergeConfigs(userConfig)
-{
-  const config = applyDefaults(userConfig);
-
-  warnIfNotReduxAction(config, 'defaultCommit');
-  warnIfNotReduxAction(config, 'defaultRollback');
-
-  // toggle experimental returned promises
-  if(!config.offlineActionTracker) {
-    config.offlineActionTracker = config.returnPromises
-      ? offlineActionTracker.withPromises
-      : offlineActionTracker.withoutPromises;
-    delete config.returnPromises;
-  }
-
-  return config
-}
 
 function configStore(store, config)
 {
